backend/requirements.txt
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
cssselect==1.1.0
Flask==1.1.2
Flask-MySQLdb==0.2.0
Flask-SQLAlchemy==2.4.4
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
lxml==4.6.2
MarkupSafe==1.1.1
mysqlclient==2.0.3
pyquery==1.4.3
requests==2.25.1
SQLAlchemy==1.3.23
tensorflow==2.3.0
urllib3==1.26.3
Werkzeug==1.0.1

nltk~=3.5

numpy~=1.18.5
Pillow~=8.1.2
